refactor: make core increment use bitwise operators + some fixes and improvements
line 351
> - int y = -64
> + int y = bot.world.minY;
this makes the core never resize when in worlds like the end or nether (which has the min y level as 0)

i'm also limiting the maximum resizing Y level now to the world's maximum y level so it won't overflow

also for some reason when i switch to flatlands using `/world 3` the server doesn't send the current dimension's data (like `min_y` and `height`), so i think the server sends them before at like, login? or are we meant to default to overworld's values?